fix(sdk-ui-dashboard): consolidate ToastMessages
Use the same Messages collection for ToastMessages and
DrillValidationToastMessages. This makes the DOM simpler and both types
of messages now show side by side properly.

JIRA: RAIL-4740
